codedx-analysis/src/requirements.txt

git+https://github.com/broadinstitute/dsp-appsec-codedx-api-client-python.git@b0bea14ef9f39df4e3abfb187ac7d2478ef8c865
google-api-core==1.21.0
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-core==0.28.1
google-cloud-pubsub==1.6.0
google-cloud-storage==1.13.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.30.0
httplib2==0.18.1
idna==2.9
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
urllib3==1.25.9
slackclient==2.8.0